move out src/test to test
