+++ .gitignore
@@ -109,6 +109,7 @@ yarn.lock
 
 # PRIVACY XDDDD
 src/test/config.ts
+test/config.ts
 .vscode
 
 # macOS is shit xD
