+++ test/chunk.ts
@@ -1,4 +1,4 @@
-import { Client, Intents } from '../../mod.ts'
+import { Client, Intents } from '../mod.ts'
 import { TOKEN } from './config.ts'
 
 const client = new Client()
