+++ test/debug.ts
@@ -1,11 +1,11 @@
-import { Client, event } from '../../mod.ts'
+import { Client, event } from '../mod.ts'
 import { TOKEN } from './config.ts'
 
 class MyClient extends Client {
   constructor() {
     super({
       token: TOKEN,
-      intents: [],
+      intents: []
     })
   }
 
@@ -17,7 +17,9 @@ class MyClient extends Client {
   debug(title: string, msg: string): void {
     console.log(`[${title}] ${msg}`)
     if (title === 'Gateway' && msg === 'Initializing WebSocket...') {
-      try { throw new Error("Stack") } catch (e) {
+      try {
+        throw new Error('Stack')
+      } catch (e) {
         console.log(e.stack)
       }
     }
