+++ test/guild.ts
@@ -5,7 +5,7 @@ import {
   // PermissionFlags,
   // ChannelTypes,
   // GuildCreateOptions
-} from '../../mod.ts'
+} from '../mod.ts'
 import { TOKEN } from './config.ts'
 
 const client = new Client()
